workflow doesnt drop database after tests

+++ test/test_get_new_data_from_database.sh
@@ -38,4 +38,4 @@ psql -c 'CREATE DATABASE test_database;'
 psql -f test_database.sql
 rm test_database.sql
 pytest test/test_get_new_data_from_database.py -vvvrP
-psql -c "DROP DATABASE IF EXISTS test_database"
+if  [[ $PGUSER!='postgres_user' ]]; then psql -c "DROP DATABASE IF EXISTS test_database"; fi
